Document the issue #1037
I also consider this a bit unexpected behavior. I think ideally the Gorm should see there are more `primary_key` fields in a struct and turn off `auto_increment` for them automatically. And keep `auto_increment` on by default only for the case when there is a single `primary_key` in the struct.

That may be difficult to implement, though. So as long as the current implementation is kept, the documentation should be there to clarify this behaviour.