Change DB.Where to db.Where (#549)
I believe this example "Join with conditions" query is meant to use the regular gormDB db.Where, but the docs currently have it using DB.Where

On the other hand, maybe you intended to guide users to use the regular sql db accessible at db.DB(). In that case, it should be made clear in the text preceding this example.