Update many_to_many.md (#667)
When we migrate the structs to database, the plural form of structs is used as a table name. So, the name of the mapping table of Person and Address should also be plural. Else, SetupJoinTable would not work on it and and some other operation on PersonAddress (like using its composite primary keys as foreign keys in another table) will result in new table creation which would be person_addresses.

If we name our mapping table as person_address(plural form) then it update the same table instead of creating the new one.

Co-authored-by: Muhammad Amir Ejaz <37077032+BCSF15M004@users.noreply.github.com>