Changed the first example in "Embedded Struct"
In the first example, earlier gorm.Model was used to depict an embedded struct. For new users like myself, there might be confusion on what gorm.Model exactly is, and why we have done it so. Hence, here, I depict the usage through another struct within the example (structs copied from the example below this one)