Grammatical improvement query.md
As for the first change, switched the word order on line 36 to avoid ambiguity in "is not performant  and nondeterministic" (to "is non0deterministic and not performant"). The second change is the preferable spelling of" nonperformant"->"non-performant". The  spelling is listed first in the Cambridge dictionary(dictionary.cambridge.org/dictionary/english/non-deterministic)